adds tests for PathUtils createHref function with options to keep trailing slash when combining basename with root location.pathname.
moves all tests for PathUtils createHref to separate
file (PathUtils-createHref-test.js)
